+++ core/src/main/scala/kafka/server/AbstractFetcherManager.scala
@@ -18,10 +18,10 @@
 package kafka.server
 
 import kafka.utils.Logging
-import org.apache.kafka.common.network.BrokerEndPoint
 import org.apache.kafka.common.{TopicPartition, Uuid}
 import org.apache.kafka.common.utils.Utils
 import org.apache.kafka.server.metrics.KafkaMetricsGroup
+import org.apache.kafka.server.network.BrokerEndPoint
 
 import scala.collection.{Map, Set, mutable}
 import scala.jdk.CollectionConverters._
