+++ core/src/main/scala/kafka/server/LeaderEndPoint.scala
@@ -22,8 +22,8 @@ import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.requests.FetchRequest
 import org.apache.kafka.common.message.OffsetForLeaderEpochResponseData.EpochEndOffset
 import org.apache.kafka.common.message.{FetchResponseData, OffsetForLeaderEpochRequestData}
-import org.apache.kafka.common.network.BrokerEndPoint
 import org.apache.kafka.server.common.OffsetAndEpoch
+import org.apache.kafka.server.network.BrokerEndPoint
 
 import scala.collection.Map
