+++ core/src/main/scala/kafka/server/ReplicaAlterLogDirsManager.scala
@@ -18,8 +18,8 @@
 package kafka.server
 
 import org.apache.kafka.common.TopicPartition
-import org.apache.kafka.common.network.BrokerEndPoint
 import org.apache.kafka.server.common.DirectoryEventHandler
+import org.apache.kafka.server.network.BrokerEndPoint
 import org.apache.kafka.storage.log.metrics.BrokerTopicStats
 
 class ReplicaAlterLogDirsManager(brokerConfig: KafkaConfig,
