+++ core/src/main/scala/kafka/server/ReplicaFetcherManager.scala
@@ -19,9 +19,9 @@ package kafka.server
 
 import org.apache.kafka.clients.FetchSessionHandler
 import org.apache.kafka.common.metrics.Metrics
-import org.apache.kafka.common.network.BrokerEndPoint
 import org.apache.kafka.common.utils.{LogContext, Time}
 import org.apache.kafka.server.common.MetadataVersion
+import org.apache.kafka.server.network.BrokerEndPoint
 
 class ReplicaFetcherManager(brokerConfig: KafkaConfig,
                             protected val replicaManager: ReplicaManager,
