+++ core/src/test/scala/unit/kafka/server/FetcherThreadTestUtils.scala
@@ -19,8 +19,8 @@ package kafka.server
 
 import org.apache.kafka.common.Uuid
 import org.apache.kafka.common.compress.Compression
-import org.apache.kafka.common.network.BrokerEndPoint
 import org.apache.kafka.common.record.{MemoryRecords, RecordBatch, SimpleRecord}
+import org.apache.kafka.server.network.BrokerEndPoint
 
 import scala.jdk.CollectionConverters._
