+++ server-common/src/main/java/org/apache/kafka/server/network/BrokerEndPoint.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.network;
+package org.apache.kafka.server.network;
 
 import java.util.Objects;
